[Refactor] 통합 및 config 클래스 위치 변경
Member에 있던 Config 클래스들 패키지 변경 및 Swagger-ui 적용
