+++ src/main/java/com/woof/api/config/SwaggerConfig.java
@@ -1,0 +1,30 @@
+package com.woof.api.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+import java.util.ArrayList;
+
+@Configuration
+public class SwaggerConfig {
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(
+                        RequestHandlerSelectors.basePackage("com.woof.api")
+                )
+                .paths(PathSelectors.any())
+                .build();
+    }
+
+
+
+
+}
