+++ src/main/java/com/woof/api/productCeo/repository/ProductCeoRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface ProductCeoRepository extends JpaRepository<ProductCeo, Long>
 //        , ProductRepositoryCustum
 {
-    public Optional<ProductCeo> findByName(String name);
+    public Optional<ProductCeo> findByStoreName(String storeName);
 
 
 //    @Query("SELECT p FROM ProductManager p " +
